build: 🛠 fix Windows build and address compiler warnings
- Create ZLIB::ZLIB alias for compatibility with find_package(ZLIB)
- Explicitly link zlib to OpenXLSX to satisfy transitive dependencies
- Fix C4244 and C4267 data loss warnings in testXLRow.cpp and XLRowData.hpp by using static_cast and float literals

+++ Tests/testXLRow.cpp
@@ -32,17 +32,17 @@ TEST_CASE("XLRow Tests", "[XLRow]")
 
         XLRow copy3;
         copy3 = copy2;
-        copy3.setHeight(height * 3);
-        REQUIRE(copy3.height() == height * 3);
-        copy3.setHeight(height * 4);
-        REQUIRE(copy3.height() == height * 4);
+        copy3.setHeight(height * 3.0f);
+        REQUIRE(copy3.height() == height * 3.0f);
+        copy3.setHeight(height * 4.0f);
+        REQUIRE(copy3.height() == height * 4.0f);
 
         XLRow copy4;
         copy4 = std::move(copy3);
-        copy4.setDescent(descent * 2);
-        REQUIRE(copy4.descent() == descent * 2);
-        copy4.setDescent(descent * 3);
-        REQUIRE(copy4.descent() == descent * 3);
+        copy4.setDescent(descent * 2.0f);
+        REQUIRE(copy4.descent() == descent * 2.0f);
+        copy4.setDescent(descent * 3.0f);
+        REQUIRE(copy4.descent() == descent * 3.0f);
 
         REQUIRE(copy4.isHidden() == false);
         copy4.setHidden(true);
@@ -166,12 +166,12 @@ TEST_CASE("XLRowData Tests", "[XLRowData]")
 
         auto       val1sum      = 0;
         const auto val1results1 = static_cast<std::vector<std::string>>(row1c.values());
-        for (const auto v : val1results1) val1sum += v.size();
+        for (const auto v : val1results1) val1sum += static_cast<int>(v.size());
         REQUIRE(val1sum == 12);
 
         val1sum                 = 0;
         const auto val1results2 = static_cast<std::vector<XLCellValue>>(row1c.values());
-        for (const auto v : val1results2) val1sum += v.get<std::string>().size();
+        for (const auto v : val1results2) val1sum += static_cast<int>(v.get<std::string>().size());
         REQUIRE(val1sum == 12);
 
         row1.values().clear();
@@ -303,12 +303,12 @@ TEST_CASE("XLRowData Tests", "[XLRowData]")
 
         auto       val1sum      = 0;
         const auto val1results1 = static_cast<std::list<std::string>>(row1c.values());
-        for (const auto v : val1results1) val1sum += v.size();
+        for (const auto v : val1results1) val1sum += static_cast<int>(v.size());
         REQUIRE(val1sum == 12);
 
         val1sum                 = 0;
         const auto val1results2 = static_cast<std::list<XLCellValue>>(row1c.values());
-        for (const auto v : val1results2) val1sum += v.get<std::string>().size();
+        for (const auto v : val1results2) val1sum += static_cast<int>(v.get<std::string>().size());
         REQUIRE(val1sum == 12);
 
         row1.values().clear();
@@ -440,12 +440,12 @@ TEST_CASE("XLRowData Tests", "[XLRowData]")
 
         auto       val1sum      = 0;
         const auto val1results1 = static_cast<std::deque<std::string>>(row1c.values());
-        for (const auto v : val1results1) val1sum += v.size();
+        for (const auto v : val1results1) val1sum += static_cast<int>(v.size());
         REQUIRE(val1sum == 12);
 
         val1sum                 = 0;
         const auto val1results2 = static_cast<std::deque<XLCellValue>>(row1c.values());
-        for (const auto v : val1results2) val1sum += v.get<std::string>().size();
+        for (const auto v : val1results2) val1sum += static_cast<int>(v.get<std::string>().size());
         REQUIRE(val1sum == 12);
 
         row1.values().clear();
